UI fixed, Car Input Fixed, Character Input Fixed, Spawning In on character fixed too
We fixed everything regarding controlling the car and character. You can now get inside of the car by pressing E and being in range which will be checked inside of the blueprints. If you are not in range you cannot get into the car but you can get out of it. I need to still put comments on my code so I'll know where to find stuff. UI has also been fixed and only appears if your inside of the car