Fixed Camera and turned a setting off inside of the sandboxcharacter
Made the camera work with a reference from the sandbox character to the spring arm and attached that to a move compenent node. I also had to turn off a function called "use pawn control rotation" inside of the spring arm component that was located inside of the sandbox character. After that I used 2 move compenent nodes when pause is pressed once to replicate the same feeling when pressing pause in GTA and only one move component node for returning back to your character with the camera.